+++ core/tauri-runtime/src/window.rs
@@ -24,6 +24,8 @@ use std::{
 type UriSchemeProtocol =
   dyn Fn(&HttpRequest) -> Result<HttpResponse, Box<dyn std::error::Error>> + Send + Sync + 'static;
 
+type WebResourceRequestHandler = dyn Fn(&HttpRequest, &mut HttpResponse) + Send + Sync;
+
 /// UI scaling utilities.
 pub mod dpi;
 
@@ -182,7 +184,7 @@ impl<'de> Deserialize<'de> for CursorIcon {
       "grab" => CursorIcon::Grab,
       "grabbing" => CursorIcon::Grabbing,
       "allscroll" => CursorIcon::AllScroll,
-      "zoomun" => CursorIcon::ZoomIn,
+      "zoomin" => CursorIcon::ZoomIn,
       "zoomout" => CursorIcon::ZoomOut,
       "eresize" => CursorIcon::EResize,
       "nresize" => CursorIcon::NResize,
@@ -234,8 +236,10 @@ pub struct PendingWindow<T: UserEvent, R: Runtime<T>> {
   /// A handler to decide if incoming url is allowed to navigate.
   pub navigation_handler: Option<Box<dyn Fn(Url) -> bool + Send>>,
 
-  /// The current webview URL.
-  pub current_url: Arc<Mutex<Url>>,
+  pub web_resource_request_handler: Option<Box<WebResourceRequestHandler>>,
+
+  /// The resolved URL to load on the webview.
+  pub url: String,
 }
 
 pub fn is_label_valid(label: &str) -> bool {
@@ -275,7 +279,8 @@ impl<T: UserEvent, R: Runtime<T>> PendingWindow<T, R> {
         menu_ids: Arc::new(Mutex::new(menu_ids)),
         js_event_listeners: Default::default(),
         navigation_handler: Default::default(),
-        current_url: Arc::new(Mutex::new("tauri://localhost".parse().unwrap())),
+        web_resource_request_handler: Default::default(),
+        url: "tauri://localhost".to_string(),
       })
     }
   }
@@ -305,7 +310,8 @@ impl<T: UserEvent, R: Runtime<T>> PendingWindow<T, R> {
         menu_ids: Arc::new(Mutex::new(menu_ids)),
         js_event_listeners: Default::default(),
         navigation_handler: Default::default(),
-        current_url: Arc::new(Mutex::new("tauri://localhost".parse().unwrap())),
+        web_resource_request_handler: Default::default(),
+        url: "tauri://localhost".to_string(),
       })
     }
   }
@@ -346,9 +352,6 @@ pub struct JsEventListenerKey {
 /// A webview window that is not yet managed by Tauri.
 #[derive(Debug)]
 pub struct DetachedWindow<T: UserEvent, R: Runtime<T>> {
-  /// The current webview URL.
-  pub current_url: Arc<Mutex<Url>>,
-
   /// Name of the window
   pub label: String,
 
@@ -365,7 +368,6 @@ pub struct DetachedWindow<T: UserEvent, R: Runtime<T>> {
 impl<T: UserEvent, R: Runtime<T>> Clone for DetachedWindow<T, R> {
   fn clone(&self) -> Self {
     Self {
-      current_url: self.current_url.clone(),
       label: self.label.clone(),
       dispatcher: self.dispatcher.clone(),
       menu_ids: self.menu_ids.clone(),