+++ tooling/bundler/src/bundle/windows/nsis.rs
@@ -8,9 +8,9 @@ use crate::{
   bundle::{
     common::CommandExt,
     windows::util::{
-      download, download_and_verify, extract_zip, remove_unc_lossy, HashAlgorithm,
-      NSIS_OUTPUT_FOLDER_NAME, NSIS_UPDATER_OUTPUT_FOLDER_NAME, WEBVIEW2_BOOTSTRAPPER_URL,
-      WEBVIEW2_X64_INSTALLER_GUID, WEBVIEW2_X86_INSTALLER_GUID,
+      download, download_and_verify, extract_zip, HashAlgorithm, NSIS_OUTPUT_FOLDER_NAME,
+      NSIS_UPDATER_OUTPUT_FOLDER_NAME, WEBVIEW2_BOOTSTRAPPER_URL, WEBVIEW2_X64_INSTALLER_GUID,
+      WEBVIEW2_X86_INSTALLER_GUID,
     },
   },
   Settings,
@@ -26,7 +26,7 @@ use tauri_utils::{
 };
 
 use std::{
-  collections::BTreeMap,
+  collections::{BTreeMap, HashMap},
   fs::{copy, create_dir_all, remove_dir_all, rename, write},
   path::{Path, PathBuf},
   process::Command,
@@ -121,6 +121,25 @@ fn get_and_extract_nsis(nsis_toolset_path: &Path, _tauri_tools_path: &Path) -> c
   Ok(())
 }
 
+fn add_build_number_if_needed(version_str: &str) -> anyhow::Result<String> {
+  let version = semver::Version::parse(version_str).context("invalid app version")?;
+  if !version.build.is_empty() {
+    let build = version.build.parse::<u64>();
+    if build.is_ok() {
+      return Ok(format!(
+        "{}.{}.{}.{}",
+        version.major, version.minor, version.patch, version.build
+      ));
+    } else {
+      anyhow::bail!("optional build metadata in app version must be numeric-only");
+    }
+  }
+
+  Ok(format!(
+    "{}.{}.{}.0",
+    version.major, version.minor, version.patch,
+  ))
+}
 fn build_nsis_app_installer(
   settings: &Settings,
   _nsis_toolset_path: &Path,
@@ -164,7 +183,9 @@ fn build_nsis_app_installer(
   let mut data = BTreeMap::new();
 
   let bundle_id = settings.bundle_identifier();
-  let manufacturer = bundle_id.split('.').nth(1).unwrap_or(bundle_id);
+  let manufacturer = settings
+    .publisher()
+    .unwrap_or_else(|| bundle_id.split('.').nth(1).unwrap_or(bundle_id));
 
   #[cfg(not(target_os = "windows"))]
   {
@@ -177,9 +198,15 @@ fn build_nsis_app_installer(
   data.insert("bundle_id", to_json(bundle_id));
   data.insert("manufacturer", to_json(manufacturer));
   data.insert("product_name", to_json(settings.product_name()));
+  data.insert("short_description", to_json(settings.short_description()));
+  data.insert("copyright", to_json(settings.copyright_string()));
 
   let version = settings.version_string();
   data.insert("version", to_json(version));
+  data.insert(
+    "version_with_build",
+    to_json(add_build_number_if_needed(version)?),
+  );
 
   data.insert(
     "allow_downgrades",
@@ -188,34 +215,32 @@ fn build_nsis_app_installer(
 
   let mut install_mode = NSISInstallerMode::CurrentUser;
   let mut languages = vec!["English".into()];
+  let mut custom_template_path = None;
+  let mut custom_language_files = None;
   if let Some(nsis) = &settings.windows().nsis {
+    custom_template_path = nsis.template.clone();
+    custom_language_files = nsis.custom_language_files.clone();
     install_mode = nsis.install_mode;
     if let Some(langs) = &nsis.languages {
       languages.clear();
       languages.extend_from_slice(langs);
     }
     if let Some(license) = &nsis.license {
-      data.insert(
-        "license",
-        to_json(remove_unc_lossy(license.canonicalize()?)),
-      );
+      data.insert("license", to_json(dunce::canonicalize(license)?));
     }
     if let Some(installer_icon) = &nsis.installer_icon {
       data.insert(
         "installer_icon",
-        to_json(remove_unc_lossy(installer_icon.canonicalize()?)),
+        to_json(dunce::canonicalize(installer_icon)?),
       );
     }
     if let Some(header_image) = &nsis.header_image {
-      data.insert(
-        "header_image",
-        to_json(remove_unc_lossy(header_image.canonicalize()?)),
-      );
+      data.insert("header_image", to_json(dunce::canonicalize(header_image)?));
     }
     if let Some(sidebar_image) = &nsis.sidebar_image {
       data.insert(
         "sidebar_image",
-        to_json(remove_unc_lossy(sidebar_image.canonicalize()?)),
+        to_json(dunce::canonicalize(sidebar_image)?),
       );
     }
 
@@ -232,7 +257,26 @@ fn build_nsis_app_installer(
       NSISInstallerMode::Both => "both",
     }),
   );
+
+  let mut languages_data = Vec::new();
+  for lang in &languages {
+    if let Some(data) = get_lang_data(lang, custom_language_files.as_ref())? {
+      languages_data.push(data);
+    } else {
+      log::warn!("Custom tauri messages for {lang} are not translated.\nIf it is a valid language listed on <https://github.com/kichik/nsis/tree/9465c08046f00ccb6eda985abbdbf52c275c6c4d/Contrib/Language%20files>, please open a Tauri feature request\n or you can provide a custom language file for it in `tauri.conf.json > tauri > bundle > windows > nsis > custom_language_files`");
+    }
+  }
+
   data.insert("languages", to_json(languages.clone()));
+  data.insert(
+    "language_files",
+    to_json(
+      languages_data
+        .iter()
+        .map(|d| d.0.clone())
+        .collect::<Vec<_>>(),
+    ),
+  );
 
   let main_binary = settings
     .binaries()
@@ -359,29 +403,26 @@ fn build_nsis_app_installer(
     }
     output
   });
-  handlebars
-    .register_template_string("installer.nsi", include_str!("./templates/installer.nsi"))
-    .map_err(|e| e.to_string())
-    .expect("Failed to setup handlebar template");
+  if let Some(path) = custom_template_path {
+    handlebars
+      .register_template_string("installer.nsi", std::fs::read_to_string(path)?)
+      .map_err(|e| e.to_string())
+      .expect("Failed to setup custom handlebar template");
+  } else {
+    handlebars
+      .register_template_string("installer.nsi", include_str!("./templates/installer.nsi"))
+      .map_err(|e| e.to_string())
+      .expect("Failed to setup handlebar template");
+  }
   let installer_nsi_path = output_path.join("installer.nsi");
-  write(
+  write_ut16_le_with_bom(
     &installer_nsi_path,
-    encoding_rs::UTF_8
-      .encode(handlebars.render("installer.nsi", &data)?.as_str())
-      .0,
+    handlebars.render("installer.nsi", &data)?.as_str(),
   )?;
 
-  for lang in languages {
-    if let Some((data, encoding)) = get_lang_data(&lang) {
-      write(
-        output_path.join(lang).with_extension("nsh"),
-        encoding.encode(data).0,
-      )?;
-    } else {
-      return Err(
-        anyhow::anyhow!("Language {lang} not implemented. If it is a valid language listed on <https://github.com/kichik/nsis/tree/9465c08046f00ccb6eda985abbdbf52c275c6c4d/Contrib/Language%20files>, please open a Tauri feature request")
-          .into()
-      );
+  for (lang, data) in languages_data.iter() {
+    if let Some(content) = data {
+      write_ut16_le_with_bom(output_path.join(lang).with_extension("nsh"), content)?;
     }
   }
 
@@ -442,7 +483,7 @@ fn generate_resource_data(settings: &Settings) -> crate::Result<ResourcesMap> {
 
   for src in settings.resource_files() {
     let src = src?;
-    let resource_path = remove_unc_lossy(cwd.join(&src).canonicalize()?);
+    let resource_path = dunce::canonicalize(cwd.join(&src))?;
 
     // In some glob resource paths like `assets/**/*` a file might appear twice
     // because the `tauri_utils::resources::ResourcePaths` iterator also reads a directory
@@ -476,7 +517,7 @@ fn generate_binaries_data(settings: &Settings) -> crate::Result<BinariesMap> {
 
   for src in settings.external_binaries() {
     let src = src?;
-    let binary_path = remove_unc_lossy(cwd.join(&src).canonicalize()?);
+    let binary_path = dunce::canonicalize(cwd.join(&src))?;
     let dest_filename = src
       .file_name()
       .expect("failed to extract external binary filename")
@@ -502,30 +543,47 @@ fn generate_binaries_data(settings: &Settings) -> crate::Result<BinariesMap> {
   Ok(binaries)
 }
 
-fn get_lang_data(lang: &str) -> Option<(&'static str, &'static encoding_rs::Encoding)> {
-  use encoding_rs::*;
-  match lang.to_lowercase().as_str() {
-    "arabic" => Some((
-      include_str!("./templates/nsis-languages/Arabic.nsh"),
-      UTF_16LE,
-    )),
-    "english" => Some((
-      include_str!("./templates/nsis-languages/English.nsh"),
-      UTF_8,
-    )),
-    "portuguesebr" => Some((
-      include_str!("./templates/nsis-languages/PortugueseBR.nsh"),
-      UTF_8,
-    )),
-    "tradchinese" => Some((
-      include_str!("./templates/nsis-languages/TradChinese.nsh"),
-      UTF_8,
-    )),
-    "simpchinese" => Some((
-      include_str!("./templates/nsis-languages/SimpChinese.nsh"),
-      UTF_8,
-    )),
-    "french" => Some((include_str!("./templates/nsis-languages/French.nsh"), UTF_8)),
-    _ => None,
+fn get_lang_data(
+  lang: &str,
+  custom_lang_files: Option<&HashMap<String, PathBuf>>,
+) -> crate::Result<Option<(PathBuf, Option<&'static str>)>> {
+  if let Some(path) = custom_lang_files.and_then(|h| h.get(lang)) {
+    return Ok(Some((dunce::canonicalize(path)?, None)));
   }
+
+  let lang_path = PathBuf::from(format!("{lang}.nsh"));
+  let lang_content = match lang.to_lowercase().as_str() {
+    "arabic" => Some(include_str!("./templates/nsis-languages/Arabic.nsh")),
+    "dutch" => Some(include_str!("./templates/nsis-languages/Dutch.nsh")),
+    "english" => Some(include_str!("./templates/nsis-languages/English.nsh")),
+    "japanese" => Some(include_str!("./templates/nsis-languages/Japanese.nsh")),
+    "korean" => Some(include_str!("./templates/nsis-languages/Korean.nsh")),
+    "portuguesebr" => Some(include_str!("./templates/nsis-languages/PortugueseBR.nsh")),
+    "tradchinese" => Some(include_str!("./templates/nsis-languages/TradChinese.nsh")),
+    "simpchinese" => Some(include_str!("./templates/nsis-languages/SimpChinese.nsh")),
+    "french" => Some(include_str!("./templates/nsis-languages/French.nsh")),
+    "spanish" => Some(include_str!("./templates/nsis-languages/Spanish.nsh")),
+    "spanishinternational" => Some(include_str!(
+      "./templates/nsis-languages/SpanishInternational.nsh"
+    )),
+    "persian" => Some(include_str!("./templates/nsis-languages/Persian.nsh")),
+    "turkish" => Some(include_str!("./templates/nsis-languages/Turkish.nsh")),
+    "swedish" => Some(include_str!("./templates/nsis-languages/Swedish.nsh")),
+    _ => return Ok(None),
+  };
+
+  Ok(Some((lang_path, lang_content)))
+}
+
+fn write_ut16_le_with_bom<P: AsRef<Path>>(path: P, content: &str) -> crate::Result<()> {
+  use std::fs::File;
+  use std::io::{BufWriter, Write};
+
+  let file = File::create(path)?;
+  let mut output = BufWriter::new(file);
+  output.write_all(&[0xFF, 0xFE])?; // the BOM part
+  for utf16 in content.encode_utf16() {
+    output.write_all(&utf16.to_le_bytes())?;
+  }
+  Ok(())
 }